refactor : LottoMain 의 뷰와 모델을 제어하는 책임 Service 계층을 추가해 Controller에 위임
