+++ src/main/java/lotto/config/LottoFactory.java
@@ -7,20 +7,17 @@ import lotto.domain.LottoRandomNumberStrategy;
 import lotto.domain.ManualLottoMachine;
 import lotto.domain.RandomLottoGenerator;
 import lotto.domain.RandomNumberStrategy;
+import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class LottoFactory {
-    public InputView inputView() {
-        return new InputView(scanner());
-    }
-
-    private Scanner scanner() {
-        return new Scanner(System.in);
-    }
-
     public LottoController lottoController() {
-        return new LottoController(manualLottoMachine(), automaticLottoMachine());
+        return new LottoController(lottoService(), inputView(), outputView());
+    }
+
+    private LottoService lottoService() {
+        return new LottoService(manualLottoMachine(), automaticLottoMachine());
     }
 
     private ManualLottoMachine manualLottoMachine() {
@@ -39,7 +36,15 @@ public class LottoFactory {
         return new LottoRandomNumberStrategy();
     }
 
-    public OutputView outputView() {
+    private InputView inputView() {
+        return new InputView(scanner());
+    }
+
+    private Scanner scanner() {
+        return new Scanner(System.in);
+    }
+
+    private OutputView outputView() {
         return new OutputView();
     }
 }
